Revert "changes for query"
