+++ src/main/resources/application.properties
@@ -3,7 +3,7 @@ spring.datasource.url=jdbc:mysql://localhost:3306/pricepegging
 spring.datasource.jdbcUrl=jdbc:mysql://localhost:3306/pricepegging
 
 spring.datasource.username = root
-spring.datasource.password =root
+spring.datasource.password =Weldone@3210
 #spring.datasource.username = abhinay
 #spring.datasource.password = P@$$w0rd
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
